pyproject.toml
[project]
name = "pollen-history-collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx",
  "aiofiles",
  "pillow"
]
